Show using floating point input

+++ pybind11/Pet.cpp
@@ -30,12 +30,12 @@ py::array Pet::nump(py::array arr) {
 
     auto arr_obj_prop = arr.request();
     //initialize values
-    double *vals = (double*) arr_obj_prop.ptr;
+    float *vals = (float*) arr_obj_prop.ptr;
 
     unsigned int shape_1 = arr_obj_prop.shape[0];
     unsigned int shape_2 = arr_obj_prop.shape[1];
 
-    std::vector<std::vector <double>> vect_arr( shape_1, std::vector<double> (shape_2));
+    std::vector<std::vector <float>> vect_arr( shape_1, std::vector<float> (shape_2));
 
     for(unsigned int i = 0; i < shape_1; i++){
       for(unsigned int j = 0; j < shape_2; j++){
